Fixed Pizza.py to be apt replacement for fasterPizza.py
FasterPizza.py has a better structure for plotting data and returning data for other fuctions I changed Pizza.py to have this. Just like in FasterPizza.py I added a separate drawing function, the ability to change n in function call, and the returning of line angle data that allows drawing. This is just in case Grinberg really had his hear set on the first method of doing this